changed double to float

+++ lab4/parallel.cpp
@@ -3,21 +3,21 @@
 #include <iostream>
 #include <cmath>
 
-const double a = 1e2; // параметр уравнения
-const double epsilon = 1e-8; // порог сходимости
-const double D_x = 2.0, D_y = 2.0, D_z = 2.0; // область моделирования: [-1, 1] x [-1, 1] x [-1, 1]
-const double x_0 = -1.0, y_0 = -1.0, z_0 = -1.0;
+const float a = 1e2; // параметр уравнения
+const float epsilon = 1e-8; // порог сходимости
+const float D_x = 2.0, D_y = 2.0, D_z = 2.0; // область моделирования: [-1, 1] x [-1, 1] x [-1, 1]
+const float x_0 = -1.0, y_0 = -1.0, z_0 = -1.0;
 const int Nx = 100, Ny = 100, Nz = 100; // параметры сетки
-const double hx = D_x / (Nx - 1); // шаги сетки по различным осям (расстояния между узлами сетки)
-const double hy = D_y / (Ny - 1);
-const double hz = D_z / (Nz - 1);
+const float hx = D_x / (Nx - 1); // шаги сетки по различным осям (расстояния между узлами сетки)
+const float hy = D_y / (Ny - 1);
+const float hz = D_z / (Nz - 1);
 const int max_iterarions = 10000;
 
-double phi(double x, double y, double z) {
+float phi(float x, float y, float z) {
     return x*x + y*y + z*z;
 }
 
-double rho(double x, double y, double z) { 
+float rho(float x, float y, float z) { 
     return 6.0 - a * phi(x, y, z); 
 }
 
@@ -26,19 +26,19 @@ int idx(int i, int j, int k) {
 }
 
 void for_slice(int i, 
-               double* phi_old, double* phi_new, double* rho_vals,
-               double& max_diff, int start_x, int shadow) {
+               float* phi_old, float* phi_new, float* rho_vals,
+               float& max_diff, int start_x, int shadow) {
     for (int j = 1; j < Ny - 1; ++j) {
         for (int k = 1; k < Nz - 1; ++k) {
-            double term_x = (phi_old[idx(i+1,j,k)] + phi_old[idx(i-1,j,k)]) / (hx*hx);
-            double term_y = (phi_old[idx(i,j+1,k)] + phi_old[idx(i,j-1,k)]) / (hy*hy);
-            double term_z = (phi_old[idx(i,j,k+1)] + phi_old[idx(i,j,k-1)]) / (hz*hz);
+            float term_x = (phi_old[idx(i+1,j,k)] + phi_old[idx(i-1,j,k)]) / (hx*hx);
+            float term_y = (phi_old[idx(i,j+1,k)] + phi_old[idx(i,j-1,k)]) / (hy*hy);
+            float term_z = (phi_old[idx(i,j,k+1)] + phi_old[idx(i,j,k-1)]) / (hz*hz);
 
-            double numerator = term_x + term_y + term_z - rho_vals[idx(i,j,k)];
-            double denom = 2.0/(hx*hx) + 2.0/(hy*hy) + 2.0/(hz*hz) + a;
+            float numerator = term_x + term_y + term_z - rho_vals[idx(i,j,k)];
+            float denom = 2.0/(hx*hx) + 2.0/(hy*hy) + 2.0/(hz*hz) + a;
             phi_new[idx(i,j,k)] = numerator / denom;
 
-            double diff = std::abs(phi_new[idx(i,j,k)] - phi_old[idx(i,j,k)]);
+            float diff = std::abs(phi_new[idx(i,j,k)] - phi_old[idx(i,j,k)]);
             max_diff = std::max(max_diff, diff);
         }
     }
@@ -67,9 +67,9 @@ int main(int argc, char* argv[]) {
     int local_nx = end_x - start_x + 1 + 2*shadow; // строки которые обрабатывает текущий процесс + 2 теневых слоя
 
     // 3d массивы 
-    double* phi_old = new double[local_nx * Ny * Nz]();
-    double* phi_new = new double[local_nx * Ny * Nz]();
-    double* rho_vals = new double[local_nx * Ny * Nz]();
+    float* phi_old = new float[local_nx * Ny * Nz]();
+    float* phi_new = new float[local_nx * Ny * Nz]();
+    float* rho_vals = new float[local_nx * Ny * Nz]();
 
     // инициализация краевых узлов в 3d массивах - ЕДИНОЖДЫ
     // УСЛОВИЕ: если узел на границе, то его значение будет равно phi(x,y,z)
@@ -78,9 +78,9 @@ int main(int argc, char* argv[]) {
         int global_i = start_x + (i - shadow);
         for (int j = 0; j < Ny; ++j) {
             for (int k = 0; k < Nz; ++k) {
-                double x = x_0 + global_i * hx;
-                double y = y_0 + j * hy;
-                double z = z_0 + k * hz;
+                float x = x_0 + global_i * hx;
+                float y = y_0 + j * hy;
+                float z = z_0 + k * hz;
                 rho_vals[idx(i,j,k)] = rho(x, y, z);
                 bool boundary = (global_i == 0 || global_i == Nx - 1 || j == 0 || j == Ny - 1 || k == 0 || k == Nz - 1);
                 if (boundary) {
@@ -92,11 +92,11 @@ int main(int argc, char* argv[]) {
 
     // 3 - итерационный процесс
     int iterations = 0;
-    double max_diff, global_max_diff = epsilon + 1.0;
+    float max_diff, global_max_diff = epsilon + 1.0;
     MPI_Request requests[4];
     int request_count;
 
-    double start_time = MPI_Wtime();
+    float start_time = MPI_Wtime();
 
     do {
         request_count = 0;
@@ -105,15 +105,15 @@ int main(int argc, char* argv[]) {
         // (<-) я левому соседу отправляю мой первый реальный слой (shadow). левый сосед примет этот слой в свой правый ghost слой (local_nx-1)
         // (->) левый сосед отправляет мне свой последний реальный слой (local_nx-shadow-1) и я его пишу в свой левый ghost слой (0)
         if (rank > 0) {
-            MPI_Isend(&phi_old[shadow * Ny * Nz], Ny * Nz, MPI_DOUBLE, rank - 1, 0, MPI_COMM_WORLD, &requests[request_count++]);
-            MPI_Irecv(&phi_old[0], Ny * Nz, MPI_DOUBLE, rank - 1, 0, MPI_COMM_WORLD, &requests[request_count++]);
+            MPI_Isend(&phi_old[shadow * Ny * Nz], Ny * Nz, MPI_FLOAT, rank - 1, 0, MPI_COMM_WORLD, &requests[request_count++]);
+            MPI_Irecv(&phi_old[0], Ny * Nz, MPI_FLOAT, rank - 1, 0, MPI_COMM_WORLD, &requests[request_count++]);
         }
 
         // (->) я правому соседу отправляю мой последний реальный слой (local_nx-shadow-1). правый сосед примет его в левый ghost слой (0)
         // (<-) правый сосед отправляет мне свой первый реальный слой (shadow) и я его пишу в свой правый ghost слой (local_nx-1)
         if (rank < size - 1) {
-            MPI_Isend(&phi_old[(local_nx - shadow - 1) * Ny * Nz], Ny * Nz, MPI_DOUBLE, rank + 1, 0, MPI_COMM_WORLD, &requests[request_count++]);
-            MPI_Irecv(&phi_old[(local_nx - 1) * Ny * Nz], Ny * Nz, MPI_DOUBLE, rank + 1, 0, MPI_COMM_WORLD, &requests[request_count++]);
+            MPI_Isend(&phi_old[(local_nx - shadow - 1) * Ny * Nz], Ny * Nz, MPI_FLOAT, rank + 1, 0, MPI_COMM_WORLD, &requests[request_count++]);
+            MPI_Irecv(&phi_old[(local_nx - 1) * Ny * Nz], Ny * Nz, MPI_FLOAT, rank + 1, 0, MPI_COMM_WORLD, &requests[request_count++]);
         }
 
         // вычисление внутренних узлов (по оси X: с shadow+1 по local_nx-shadow-1; )
@@ -133,7 +133,7 @@ int main(int argc, char* argv[]) {
         for_slice(i, phi_old, phi_new, rho_vals, max_diff, start_x, shadow);
 
         // по операции MPI_MAX собирает значения max_diff со всех процессов и выдаёт global_max_diff
-        MPI_Allreduce(&max_diff, &global_max_diff, 1, MPI_DOUBLE, MPI_MAX, MPI_COMM_WORLD);
+        MPI_Allreduce(&max_diff, &global_max_diff, 1, MPI_FLOAT, MPI_MAX, MPI_COMM_WORLD);
         // if (rank == 0) std::cout << "Iteration: " << iterations << " Max diff: " << global_max_diff << std::endl;
 
         // std::swap(phi_old, phi_new);
@@ -147,7 +147,7 @@ int main(int argc, char* argv[]) {
         ++iterations;
     } while (global_max_diff > epsilon && iterations < max_iterarions);
 
-    double end_time = MPI_Wtime();
+    float end_time = MPI_Wtime();
 
     // 4 - вывод результатов и завершение MPI программы
     if (rank == 0) {
